refactor: rename graph.py and delete old tn implementation
